Completed data refactoring
