--- base/shared/src/main/scala/scalaz/data/Maybe2Syntax.scala
@@ -1,8 +1,0 @@
-package scalaz
-package data
-
-trait Maybe2Syntax {
-
-  def empty2[A, B]            = Maybe2.empty2[A, B]
-  def just2[A, B](a: A, b: B) = Maybe2.just2(a, b)
-}
