--- base/shared/src/main/scala/scalaz/data/MaybeFunctions.scala
@@ -1,7 +1,0 @@
-package scalaz
-package data
-
-trait MaybeFunctions {
-  def empty[A]      = Maybe.empty[A]
-  def just[A](a: A) = Maybe.just(a)
-}
