--- base/shared/src/main/scala/scalaz/data/UpStar.scala
@@ -1,4 +1,0 @@
-package scalaz
-package data
-
-final case class UpStar[F[_], A, B](run: A => F[B]) extends AnyVal
