+++ base/shared/src/main/scala/scalaz/data/aList1.scala
@@ -1,7 +1,6 @@
 package scalaz
 package data
 
-import Prelude._
 import AList.aListOps
 
 /**
