+++ base/shared/src/main/scala/scalaz/data/identity.scala
@@ -7,6 +7,7 @@ import scalaz.ct.MonadClass
 final case class Identity[A](run: A) extends AnyVal
 
 trait IdentityInstances {
+
   implicit val monad: Monad[Identity] = instanceOf(new MonadClass[Identity] {
     override def map[A, B](fa: Identity[A])(f: A => B): Identity[B]               = Identity(f(fa.run))
     override def ap[A, B](fa: Identity[A])(f: Identity[A => B]): Identity[B]      = Identity(f.run(fa.run))
