+++ base/shared/src/main/scala/scalaz/data/leibniz.scala
