+++ base/shared/src/main/scala/scalaz/data/liskov.scala
