+++ base/shared/src/main/scala/scalaz/data/notIs.scala
