+++ base/shared/src/main/scala/scalaz/data/void.scala
